+++ UDP/private/FileWriter.h
@@ -17,14 +17,14 @@ namespace udp
 		void* m_fHandle = nullptr;
 		jobs::Job* m_onFinished = nullptr;
 
-		size_t m_fileSize = 0;
-		size_t m_curByte = 0;
+		ull m_fileSize = 0;
+		ull m_curByte = 0;
 		bool m_busy = false;
 
 		void Write(UDPRes* resp);
 
 	public:
-		FileWriter(const std::string& path, size_t fileSize);
+		FileWriter(const std::string& path, ull fileSize);
 		~FileWriter();
 
 		void PushToQueue(UDPRes* resp);
